refactor(vms/saevm): Allow parsing before state sync
