+++ vms/saevm/cchain/statesync/handler.go
@@ -46,6 +46,7 @@ func New(
 		cfg,
 		db,
 		log,
+		hooks,
 	)
 	if err != nil {
 		return nil, fmt.Errorf("creating SAE statesync handler: %v", err)
