+++ vms/saevm/sae/BUILD.bazel
@@ -61,8 +61,6 @@ go_library(
         "@com_github_ava_labs_libevm//event",
         "@com_github_ava_labs_libevm//libevm",
         "@com_github_ava_labs_libevm//params",
-        "@com_github_ava_labs_libevm//rlp",
-        "@com_github_ava_labs_libevm//trie",
         "@com_github_ava_labs_libevm//triedb",
         "@com_github_holiman_uint256//:uint256",
         "@com_github_prometheus_client_golang//prometheus",
@@ -234,6 +232,7 @@ go_test(
         "@com_github_ava_labs_libevm//libevm/options",
         "@com_github_ava_labs_libevm//log",
         "@com_github_ava_labs_libevm//params",
+        "@com_github_ava_labs_libevm//rlp",
         "@com_github_ava_labs_libevm//rpc",
         "@com_github_ava_labs_libevm//trie",
         "@com_github_google_go_cmp//cmp",
