+++ vms/saevm/statesync/BUILD.bazel
@@ -20,6 +20,7 @@ go_library(
         "//utils/logging",
         "//vms/saevm/adaptor",
         "//vms/saevm/blocks",
+        "//vms/saevm/hook",
         "//vms/saevm/saedb",
         "@com_github_ava_labs_libevm//common",
         "@com_github_ava_labs_libevm//core/rawdb",
